backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djoser
psycopg2-binary==2.8.6


